Add Task editing functionality
