+++ app/src/main/java/com/example/tasks/ui/screen/TaskListScreen.kt
@@ -1,6 +1,7 @@
 package com.example.tasks.ui.screen
 
 import androidx.compose.animation.animateContentSize
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.material.*
@@ -9,7 +10,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.example.tasks.R
 import com.example.tasks.data.Category
@@ -23,6 +23,7 @@ sealed class TaskListEvent {
     object ShowCompleted : TaskListEvent()
     object AddTask : TaskListEvent()
     data class TaskCompleted(val id: Long) : TaskListEvent()
+    data class ShowTaskDetail(val id: Long) : TaskListEvent()
 }
 
 @Composable
@@ -114,7 +115,12 @@ fun TaskItem(task: Task, onEvent: (TaskListEvent) -> Unit) {
                 onCheckedChange = {
                     onEvent(TaskListEvent.TaskCompleted(task.taskId))
                 })
-            Text(task.taskName)
+            Text(
+                task.taskName,
+                modifier = Modifier.clickable(onClick = {
+                    onEvent(TaskListEvent.ShowTaskDetail(task.taskId))
+                })
+            )
         }
         Text(text = task.deadline?.format() ?: "No deadline", modifier = Modifier.padding(8.dp))
     }
