+++ app/src/main/res/navigation/navigation.xml
@@ -8,6 +8,9 @@
         android:id="@+id/taskListFragment"
         android:name="com.example.tasks.ui.fragment.TaskListFragment"
         android:label="TaskListFragment">
+        <action
+            android:id="@+id/action_taskListFragment_to_taskEditFragment"
+            app:destination="@id/taskEditFragment" />
     </fragment>
     <fragment
         android:id="@+id/taskAddFragment"
@@ -16,5 +19,9 @@
     <fragment
         android:id="@+id/taskEditFragment"
         android:name="com.example.tasks.ui.fragment.TaskEditFragment"
-        android:label="fragment_task_edit"/>
+        android:label="fragment_task_edit">
+        <argument
+            android:name="id"
+            app:argType="long" />
+    </fragment>
 </navigation>
